+++ src/TrashLib.Tests/Sonarr/ReleaseProfile/ReleaseProfileDataFiltererTest.cs
@@ -18,7 +18,9 @@ public class ReleaseProfileDataFiltererTest
         {
             new() {TrashId = "1", Term = "term1"},
             new() {TrashId = "2", Term = "term2"},
-            new() {TrashId = "3", Term = "term3"}
+            new() {TrashId = "3", Term = "term3"},
+            new() {Term = "term4"},
+            new() {Term = "term5"}
         };
 
         var result = sut.IncludeTerms(terms, filter);
@@ -42,7 +44,9 @@ public class ReleaseProfileDataFiltererTest
                 {
                     new() {TrashId = "1", Term = "term1"},
                     new() {TrashId = "2", Term = "term2"},
-                    new() {TrashId = "3", Term = "term3"}
+                    new() {TrashId = "3", Term = "term3"},
+                    new() {Term = "term4"},
+                    new() {Term = "term5"}
                 }
             },
             new()
@@ -77,14 +81,18 @@ public class ReleaseProfileDataFiltererTest
         {
             new() {TrashId = "1", Term = "term1"},
             new() {TrashId = "2", Term = "term2"},
-            new() {TrashId = "3", Term = "term3"}
+            new() {TrashId = "3", Term = "term3"},
+            new() {Term = "term4"},
+            new() {Term = "term5"}
         };
 
         var result = sut.ExcludeTerms(terms, filter);
 
         result.Should().BeEquivalentTo(new TermData[]
         {
-            new() {TrashId = "3", Term = "term3"}
+            new() {TrashId = "3", Term = "term3"},
+            new() {Term = "term4"},
+            new() {Term = "term5"}
         });
     }
 
@@ -101,7 +109,9 @@ public class ReleaseProfileDataFiltererTest
                 {
                     new() {TrashId = "1", Term = "term1"},
                     new() {TrashId = "2", Term = "term2"},
-                    new() {TrashId = "3", Term = "term3"}
+                    new() {TrashId = "3", Term = "term3"},
+                    new() {Term = "term4"},
+                    new() {Term = "term5"}
                 }
             },
             new()
@@ -109,7 +119,9 @@ public class ReleaseProfileDataFiltererTest
                 Score = 20,
                 Terms = new TermData[]
                 {
-                    new() {TrashId = "4", Term = "term4"}
+                    new() {TrashId = "4", Term = "term4"},
+                    new() {Term = "term6"},
+                    new() {Term = "term7"}
                 }
             }
         };
@@ -123,7 +135,9 @@ public class ReleaseProfileDataFiltererTest
                 Score = 10,
                 Terms = new TermData[]
                 {
-                    new() {TrashId = "3", Term = "term3"}
+                    new() {TrashId = "3", Term = "term3"},
+                    new() {Term = "term4"},
+                    new() {Term = "term5"}
                 }
             },
             new()
@@ -131,7 +145,9 @@ public class ReleaseProfileDataFiltererTest
                 Score = 20,
                 Terms = new TermData[]
                 {
-                    new() {TrashId = "4", Term = "term4"}
+                    new() {TrashId = "4", Term = "term4"},
+                    new() {Term = "term6"},
+                    new() {Term = "term7"}
                 }
             }
         });