+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Sonarr: Fix unexpected missing terms when using filters. (#69)
+
 ## [2.0.1] - 2022-05-19
 
 ### Fixed